fix: addschema moved to middleware folder
