+++ routes/api/contacts.js
@@ -1,25 +1,9 @@
 const express = require("express");
-const Joi = require("joi");
-
 const router = express.Router();
+
 const contacts = require("../../models/contacts");
 const { HttpError } = require("../../helpers");
-
-const addSchema = Joi.object({
-  name: Joi.string().min(2).max(40).required(),
-  email: Joi.string()
-    .email({
-      minDomainSegments: 2,
-      tlds: { allow: ["com", "net", "org", "ua"] },
-    })
-    .pattern(
-      /^([\w-]+(?:\.[\w-]+)*)@((?:[\w-]+\.)*\w[\w-]{0,66})\.([a-z]{2,6}(?:\.[a-z]{2})?)$/i
-    )
-    .required(),
-  phone: Joi.string()
-    .pattern(/^(\+?\d+)?\s*(\(\d+\))?[\s-]*([\d-]*)$/)
-    .required(),
-});
+const { addSchema } = require("../../middleware/addSchema");
 
 router.get("/", async (req, res, next) => {
   try {
